requirements.txt
# --- Core Framework (核心框架) ---
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.1
python-multipart>=0.0.9    # 处理文件上传

# --- Database & ORM (数据库) ---
sqlalchemy>=2.0.25         # ORM
alembic>=1.13.0            # 迁移工具
asyncpg>=0.29.0            # 异步 PG 驱动
psycopg2-binary>=2.9.9     # 同步 PG 驱动 (Alembic用)
greenlet>=3.0.0            # SQLAlchemy 异步依赖

# --- Security & Auth (安全与认证) ---
python-jose>=3.3.0         # JWT 处理 (核心)
cryptography>=41.0.0       # JWT 依赖 (加密库)
passlib>=1.7.4             # 密码哈希工具
bcrypt==4.0.1              # ⚠️ 锁定版本以解决 passlib 兼容性问题

# --- Async Tasks & Caching (异步任务与缓存) ---
celery>=5.3.6              # 任务队列
redis>=5.0.1               # Redis 驱动 (Cache 和 Celery 共用)
fastapi-cache2>=0.2.1      # 接口缓存

# --- SaaS Business Modules (业务模块) ---
fastapi-mail>=1.4.1        # 邮件服务
stripe>=7.0.0              # 支付 SDK
fastapi-limiter>=0.1.6     # API 限流

# --- Data Processing & Storage (数据与存储) ---
pandas>=2.1.0              # Excel 处理
openpyxl>=3.1.2            # Excel 引擎
boto3>=1.34.0              # AWS S3 SDK

# --- Logging & Monitoring (日志与监控) ---
loguru>=0.7.2              # 结构化日志

# --- Testing (测试) ---
pytest>=7.4.0
pytest-asyncio>=0.23.0     # 异步测试插件
httpx>=0.26.0              # HTTP 客户端
pytest-dotenv>=0.5.2       # 测试环境加载 .env